+++ src/main/java/product_etl/Purchase.scala
@@ -6,7 +6,7 @@ import org.json4s.jackson.Serialization.read
 /**
   * @author Evgeny Borisov
   */
-case class Purchase(product_id:Int,client_id:Int,cashbox_id:Int,price:Int=0) {
+case class Purchase(product_id:Long,client_id:Int,cashbox_id:Int,price:Int=0) {
 
 }
 object Purchase{
